BUG: downcasting via com._possibly_downcast_to_dtype now rounds before casting      to avoid precision issues on platform; dtype conversion to int is not more robust      related GH5174

+++ pandas/tests/test_generic.py
@@ -659,7 +659,7 @@ class TestDataFrame(unittest.TestCase, Generic):
         expected = df.copy()
         expected['A'].iloc[2] = 3
         expected['A'].iloc[5] = 6
-        assert_frame_equal(result, expected)
+        assert_frame_equal(result, expected.astype(np.int64))
 
         result = df.interpolate(method='krogh')
         expectedk = df.copy()
